+++ src/editor/core/plugins/mention/input-rules.js
@@ -1,4 +1,6 @@
 import {InputRule} from "prosemirror-inputrules"
+import {TextSelection} from "prosemirror-state";
+import {hasMark} from "../../util/node";
 
 // https://github.com/ProseMirror/prosemirror/issues/262
 const objectReplacementCharacter = '\ufffc';
@@ -8,8 +10,16 @@ let mentionRule = function(schema) {
         const mark = schema.mark('mentionQuery');
         const mentionText = schema.text('@', [mark]);
 
+        // Prevents an error log when using IME
+        if(hasMark(state.selection.$anchor.nodeBefore, mark)) {
+            return;
+        }
+
+        start = start + (match[0].length -1);
+
         return state.tr
             .removeMark(0, state.doc.nodeSize -2, mark)
+            .setSelection(TextSelection.create(state.doc,  start, end))
             .replaceSelectionWith(mentionText, false);
     })
 };